Add improved docs
