+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "remark-rehype",
   "version": "10.0.0",
-  "description": "remark plugin to transform to rehype",
+  "description": "remark plugin that turns markdown into HTML to support rehype",
   "license": "MIT",
   "keywords": [
     "unified",
@@ -31,6 +31,7 @@
   "main": "index.js",
   "types": "index.d.ts",
   "files": [
+    "lib/",
     "index.d.ts",
     "index.js"
   ],
@@ -38,6 +39,8 @@
     "@types/hast": "^2.0.0",
     "@types/mdast": "^3.0.0",
     "mdast-util-to-hast": "^12.0.0",
+    "rehype-raw": "^6.1.0",
+    "rehype-sanitize": "^5.0.0",
     "unified": "^10.0.0"
   },
   "devDependencies": {
@@ -56,7 +59,7 @@
     "xo": "^0.46.0"
   },
   "scripts": {
-    "build": "rimraf \"*.d.ts\" && tsc && type-coverage",
+    "build": "rimraf \"lib/**/*.d.ts\" \"*.d.ts\" && tsc && type-coverage",
     "format": "remark . -qfo && prettier . -w --loglevel warn && xo --fix",
     "test-api": "node --conditions development test.js",
     "test-coverage": "c8 --check-coverage --branches 100 --functions 100 --lines 100 --statements 100 --reporter lcov npm run test-api",
@@ -86,8 +89,9 @@
     "ignoreCatch": true,
     "#": "needed `any`s",
     "ignoreFiles": [
-      "index.d.ts",
-      "index.js"
+      "lib/index.d.ts",
+      "lib/index.js",
+      "index.d.ts"
     ]
   }
 }
